+++ crates/swc_ecma_minifier/tests/terser/compress/arrow/issue_27/output.js
@@ -1,3 +1,3 @@
-((jQuery) => {
-    jQuery("body").addClass("foo");
+((jQuery1)=>{
+    jQuery1("body").addClass("foo");
 })(jQuery);