+++ crates/swc_ecma_minifier/tests/terser/compress/functions/issue_2842/output.js
@@ -1,9 +1,13 @@
-(function () {
-    (function () {
-        if (true) {
-            const data = (function (data) {
+(function() {
+    (function() {
+        {
+            const data = function(data) {
                 return data[data[0]];
-            })([1, 2, 3]);
+            }([
+                1,
+                2,
+                3
+            ]);
             console.log(data);
         }
     })();