+++ crates/swc_ecma_minifier/tests/terser/compress/issue_2001/export_default_func_2/output.js
@@ -1,2 +1,1 @@
 export default function f() {}
-1;